refactor: user profile screen
