+++ src/palm-react-native-ui-kit/components/index.tsx
@@ -15,6 +15,7 @@ export { default as LinkExplorer } from './atoms/LinkExplorer'
 export { default as RadioIcon } from './atoms/RadioIcon'
 export { default as Row } from './atoms/Row'
 export { default as Title } from './atoms/Title'
+export { default as ViewHorizontalDivider } from './atoms/ViewHorizontalDivider'
 export { default as ChainLogoWrapper } from './molecules/ChainLogoWrapper'
 export { default as MediaRenderer } from './molecules/MediaRenderer'
 export { default as MenuItem } from './molecules/MenuItem'
